:truck: Renamed style file
