+++ src/style/loaders.js
